accept messages. ability to send

+++ src/main/java/org/artfable/telegram/api/TelegramBot.java
@@ -12,5 +12,5 @@ public interface TelegramBot {
      *
      * @see org.springframework.scheduling.annotation.Async
      */
-    void subscribeToUpdates(Integer lastId);
+    void subscribeToUpdates(Long lastId);
 }
